fix for mastering being decimal instead of int

+++ Source/Character.cs
@@ -29,7 +29,7 @@
             public class Mastering
             {
                 public string id;
-                public int progress;
+                public decimal progress;
             }
         }
